requirements.txt
Faker==18.13.0
openai==0.27.7
pymysql==1.0.3
langchain==0.0.177
streamlit==1.24.1
python-dotenv==1.0.0
openpyxl==3.1.2
altair<5
altair-viewer==0.4.0
matplotlib==3.7.2
trrex==0.0.5
scikit-learn==1.3.0
